+++ src/client/request/tests.rs
@@ -3,9 +3,7 @@ use super::*;
 #[test]
 fn test_request_new_params() {
     assert_eq!(
-        Request::new("endpoint", Some("params"))
-            .expect("request could not be constructed")
-            .url,
+        Request::new("endpoint", Some("params")).url,
         Url::parse("https://www.ruddr.io/api/workspace/endpoint?params")
             .expect("failed to parse URL")
     )
@@ -14,19 +12,15 @@ fn test_request_new_params() {
 #[test]
 fn test_request_new_empty_params() {
     assert_eq!(
-        Request::new("endpoint", None)
-            .expect("request could not be constructed")
-            .url,
+        Request::new("endpoint", None).url,
         Url::parse("https://www.ruddr.io/api/workspace/endpoint").expect("failed to parse URL")
     )
 }
 
 #[test]
-fn test_request_new_error() {
-    assert_eq!(
-        Request::new("", Some("params")).unwrap_err().to_string(),
-        "invalid empty endpoint"
-    )
+#[should_panic(expected = "endpoint must not be empty")]
+fn test_request_new_empty_endpoint_error() {
+    Request::new("", Some("params"));
 }
 
 #[tokio::test]
@@ -34,8 +28,7 @@ async fn test_request_get() {
     let client = reqwest::Client::builder()
         .build()
         .expect("client with env token could not be constructed");
-    let request = Request::new("projects/095e0780-48bf-472c-8deb-2fc3ebc7d90c", None)
-        .expect("request could not be constructed");
+    let request = Request::new("projects/095e0780-48bf-472c-8deb-2fc3ebc7d90c", None);
     let response = request
         .get(&client)
         .await